Supporting Instance Group add
Instance Group=IG
Currently Cluster is the only resource that TF uses to manage KOPS.
Cluster holds IG a member.
When Adding an IG to the Cluster, TF relates it as a resource update.
TF tries to update the resource Cluster with the new IG, which does not exist.
The fix:
In order to fix it, need to check under resource update if theIG was changed.
This is done by searching, under the kops state
that the given IG dose not exist. and adding it accordingly